+++ _src/sass/_pages.scss
@@ -1,3 +1,2 @@
 @import 'page/example';
 @import 'page/home';
-@import 'page/test';
