+++ config.js
@@ -24,7 +24,7 @@ module.exports = function() {
             data: root + '/_src/html/_data/**/*.{js,json}',
             helpers: [
                 './node_modules/handlebars-layouts',
-                root + '/_src/scripts/plugins/Helpers.js'
+                root + '/helpers.js'
             ],
             partials: [
                 root + '/_src/html/_partials/**/*.hbs',
